+++ R/class.R
@@ -47,6 +47,9 @@ setMethod(
 #' @slot data data.table of absolute copy number calling.
 #' @slot summary.per.sample data.table of copy number variation summary per sample.
 #' @slot genome_build genome build version, should be one of 'hg19' or 'hg38'.
+#' @slot genome_measure Set 'called' will use autosomo called segments size to compute total size
+#' for CNA burden calculation, this option is useful for WES and target sequencing.
+#' Set 'wg' will autosome size from genome build, this option is useful for WGS, SNP etc..
 #' @slot annotation data.table of annotation for copy number segments.
 #' @slot dropoff.segs data.table of copy number segments dropped from raw input.
 #' @slot clinical.data data associated with each sample in copy number profile.
@@ -58,6 +61,7 @@ CopyNumber = setClass(
     data = "data.table",
     summary.per.sample = 'data.table',
     genome_build = "character",
+    genome_measure = "character",
     annotation = "data.table",
     dropoff.segs = "data.table",
     clinical.data = "data.table"
@@ -100,3 +104,35 @@ setMethod(
     str(object, max.level = 2)
   }
 )
+
+
+
+# Methods -----------------------------------------------------------------
+
+#' Subsetting CopyNumber object
+#'
+#' Subset data slot of [CopyNumber] object, un-selected rows will move to
+#' dropoff.segs slot, annotation slot will update in the same way.
+#'
+#' @param x a [CopyNumber] object to be subsetted.
+#' @param subset logical expression indicating rows to keep.
+#' @param ... further arguments to be passed to or from other methods.
+#' Useless here.
+#' @author Shixiang Wang
+#' @return a [CopyNumber] object
+#' @export
+subset.CopyNumber = function(x, subset = TRUE, ...) {
+  data = x@data
+  enclos = parent.frame()
+  subset = substitute(subset)
+  row_selector = eval(subset, data, enclos)
+
+  x@dropoff.segs = rbind(x@dropoff.segs, data[!row_selector])
+  x@data = data[row_selector]
+  x@annotation = x@annotation[row_selector]
+
+  x@summary.per.sample = get_cnsummary_sample(x@data,
+                                              genome_build = x@genome_build,
+                                              genome_measure = x@genome_measure)
+  x
+}