+++ NEWS.md
@@ -1,3 +1,9 @@
+
+# sigminer 0.1.1
+
+* add `subset.CopyNumber()` for subset CopyNumber object
+* add `genome_measure` slot for CopyNumber object
+
 # sigminer 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.